cleanup and moving components to own files
