+++ frontend/components/Header/index.tsx
@@ -1,4 +1,15 @@
 import Header from "./Header";
+import NavList from "./Nav/NavList";
+import NavListItem from "./Nav/NavListItem";
+import { NavMenuButton } from "./Nav/NavMenuButton";
+import NavMenuListItem from "./Nav/NavMenuListItem";
 import NavHeader from "./NavHeader";
 
-export { Header, NavHeader };
+export {
+  Header,
+  NavHeader,
+  NavList,
+  NavMenuListItem,
+  NavListItem,
+  NavMenuButton,
+};
